requirements.txt
imutils==0.5.3
numpy==1.18.5
opencv-python==4.2.0.34
scipy==1.4.1
flask-mysqldb
Flask-Mail
matplotlib
pandas
pytest
flask-testing
